requirements.txt

azure-functions
afinn
pandas
pyodbc